requirements.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rich==14.1.0
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch-pruning==1.6.0
tqdm==4.67.1
typing_extensions==4.14.1
pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu121
pip install "numpy<2"